moved test for _LengthConsumer
